CompetitionFacade tests

+++ GamingManagement-service/src/main/java/cz/fi/muni/pa165/yellow_yak/facade/CompetitionFacadeImpl.java
@@ -29,18 +29,27 @@ public class CompetitionFacadeImpl implements CompetitionFacade {
 
     @Override
     public CompetitionDTO create(Long gameId, String name) {
+        if (gameId == null || name == null) {
+            throw new NullPointerException("arguments cannot be null");
+        }
         log.info("creating competition, gameId = {}, name = {}", gameId, name);
         return beanMappingService.mapTo(competitionService.create(gameId, name), CompetitionDTO.class);
     }
 
     @Override
     public void remove(Long id) {
+        if (id == null) {
+            throw new NullPointerException("arguments cannot be null");
+        }
         log.info("removing competition, id = {}", id);
         competitionService.remove(id);
     }
 
     @Override
     public CompetitionDTO findById(Long id) {
+        if (id == null) {
+            throw new NullPointerException("arguments cannot be null");
+        }
         log.info("finding competition by ID, id = {}", id);
         return beanMappingService.mapTo(competitionService.findById(id), CompetitionDTO.class);
     }
@@ -48,6 +57,9 @@ public class CompetitionFacadeImpl implements CompetitionFacade {
     @Override
     // TODO D1LL1G4F
     public List<CompetitionDTO> findByGame(Long gameId) {
+        if (gameId == null) {
+            throw new NullPointerException("arguments cannot be null");
+        }
         log.info("listing competition by game ID, gameId = {}", gameId);
         return beanMappingService.mapTo(competitionService.findByGame(gameId), CompetitionDTO.class);
     }
